requirements.txt
torch>=2.1.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
pyyaml>=6.0.1
einops>=0.7.0
joblib>=1.3.0
